Adding calculator page/functionality.

+++ src/main/java/com/example/demo/models/Adder.java
@@ -5,7 +5,7 @@ public class Adder {
 //	double secret2;
 	public double sum;
 	
-	public Adder (int first, double second)	{
+	public Adder (double first, double second)	{
 //		secret1 = first;
 //		secret2 = second;
 		sum = first + second;
